Use information from SoloSection for solo timing.
Instead of reading the note stream and figuring it out on the fly, TrackPlayer now uses the times from SoloSections to calculate the necessary Z values for the shader.

Provision is made for not-yet-implemented mechanisms to deal gracefully with skipping forward/back in replays and more than one solo per highway length.