Allow multiple solos to process simultaneously.
This accounts for the situation where two solos are within a couple of track lengths of each other. This design allows up to four within about two track lengths.

More specifically, the solo subshader was changed to use a pair of Float4 vectors to hold z values for up to four solos at a time. TrackMaterial was updated to assign those slots to solos as they come into range and deassign them when the solo ends so they can be reused.