Support keyframe cues, new state for searchlights
Now when keyframe cues are called for, the lights respond and swap colors as the chart calls for. Searchlights added as a special case so there would actually appear to be searchlights- ideally there'd be something a little deeper like there's gonna need to be for spotlight note support, but this is better than falling back to nothing at all.